cultural_pipeline/requirements.txt
# =======================================
# Pipeline Cultural Lima — Dependencias
# =======================================

# --- Scraping ---
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.18.0
webdriver-manager>=4.0.0

# --- Data ---
pandas>=2.2.0
numpy>=1.26.0

# --- MongoDB (Atlas sinks) ---
pymongo>=4.7.0
dnspython>=2.6.0

# --- NLP / Embeddings ---
sentence-transformers>=2.7.0   # modelo local multilingüe (recomendado)
# openai>=1.12.0               # descomenta si usas OpenAI como fallback

# --- Búsqueda vectorial ---
faiss-cpu>=1.8.0               # índice FAISS para similarity search
# faiss-gpu>=1.7.4             # usa este si tienes GPU

# --- Scheduler ---
schedule>=1.2.0

# --- Utils ---
python-dotenv>=1.0.0           # manejo de .env local
